requirement.txt
asgiref==3.7.2
channels==4.0.0
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
